ocr_server/requirements.txt
flask
opencv-python-headless
pytesseract
Pillow
numpy
